Refactor test_executions controller into multiple files
